+++ resources/views/page/show-col2.blade.php
@@ -1,0 +1,20 @@
+@extends('layouts.frontend')
+
+@section('content')
+    <form id="page-showcol2" action="/page-showpretty/{{ $context->model->id }}" method="POST">
+        @csrf
+        <x-laraknife.panels.text-area title="{{ $context->valueof('title') }}" prev="{{ $context->valueof('prev') }}"
+            up="{{ $context->valueof('up') }}"  next="{{ $context->valueof('next') }}" audio="{{ $context->valueof('audio') }}">
+            <div class="lkn-text">
+            <div class="row">
+                <div class="col-md-6">
+                {!! $context->valueOf('text1') !!}
+                </div>
+                <div class="col-md-6">
+                {!! $context->valueOf('text2') !!}
+                </div>
+            </div>
+            </div>
+        </x-laraknife.panels.text-area>
+    </form>
+@endsection